Return window even when the terminal does not support gfx
Without the window we cannot calculate the grid columns properly, but
set the column amount as one, which results the grid option not working
in terminal emulators like Alacritty that do not support Kittys's
terminal graphics protocol.

This commit removes the pattern match and returns the supports_gfx
boolean and window separately. Now when the window is returned the grid
is properly rendered.

Tested on Alacritty version `0.13.2`.